# trace: simple optimization of `span.AddEvent` operation with attributes